Lombok라이브러리 사용-getter,setter & generateDayPlans 함수 컨트롤러로 이동
