+++ travel-planner-domain/src/main/java/com/example/DayPlan.java
@@ -1,8 +1,12 @@
 package com.example;
 
+import lombok.Getter;
+import lombok.Setter;
+
 import java.time.LocalDate;
 import java.util.List;
 
+@Getter @Setter
 public class DayPlan {
     private Long id;
     private LocalDate date;
@@ -12,30 +16,6 @@ public class DayPlan {
 
     public DayPlan(LocalDate date){
         this.date = date;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public void setDate(LocalDate date) {
-        this.date = date;
-    }
-
-    public List<Location> getLocationList() {
-        return locationList;
-    }
-
-    public void setLocationList(List<Location> locationList) {
-        this.locationList = locationList;
     }
 
     @Override
