+++ travel-planner-domain/src/main/java/com/example/Location.java
@@ -1,5 +1,9 @@
 package com.example;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
 public class Location {
     private Long id;
     private String name;
@@ -11,38 +15,6 @@ public class Location {
         this.id = id;
         this.name = name;
         this.latitude = latitude;
-        this.longitude = longitude;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Double getLatitude() {
-        return latitude;
-    }
-
-    public void setLatitude(Double latitude) {
-        this.latitude = latitude;
-    }
-
-    public Double getLongitude() {
-        return longitude;
-    }
-
-    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
